Pearson {SuppDists}	R Documentation

The Pearson product moment correlation coefficient

Description

Density, distribution function, quantile function, random generator and summary function for the distribution of Pearson's product moment correlation.

Usage

dPearson(x, N, rho=0.0, log=FALSE)
pPearson(q, N, rho=0.0, lower.tail=TRUE, log.p=FALSE)
qPearson(p, N, rho=0.0, lower.tail=TRUE, log.p=FALSE)
rPearson(n, N, rho=0.0)
sPearson(N, rho=0.0)

Arguments

x,q
vector of sample correlations

p
vector of probabilities

rho
vector of population correlations

N
vector of numbers of observations, (N > 3)

n
number of values to generate. If n is a vector, length(n) values will be generated

log, log.p
logical vector; if TRUE, probabilities p are given as log(p)

lower.tail
logical vector; if TRUE (default), probabilities are P[R <= r], otherwise, P[R > r]

Value

The output values conform to the output from other such functions in R. dPearson() gives the density, pPearson() the distribution function and qPearson() its inverse. rPearson() generates random numbers. sPearson() produces a list containing parameters corresponding to the arguments – mean, median, mode, variance, sd, third cental moment, fourth central moment, Pearson's skewness, skewness, and kurtosis.

Author(s)

Bob Wheeler

Examples

pPearson(0.5, N=10)
pPearson(q=0.5, N=10, rho=0.3) 
sPearson(N=10)
plot(function(x)dPearson(x,N=10,rho=0.7),-1,1)

[Package SuppDists version 1.1-9.9 Index]