SuppDists-defunct {SuppDists}	R Documentation

Defunct Functions in Package SuppDists

Description

The functions listed here are no longer part of SuppDists as they do not work on 64bit architectures.

Usage

# Defunct in version 1.1-9.5
rziggurat(n, normal=TRUE, new.start=FALSE, seed=556677)
rMWC1019(n, new.start=FALSE, seed=556677)

Details

Both RNG functions were developed for 32Bit architecture and do not work correctly on 64Bit machines. Furthermore, the underlying C++ code produces memory leaks as detected by R CMD check --use-valgrind.

rziggurat can be substituted with zrnorm of the package RcppZiggurat.

See Also

Defunct, zrnorm

[Package SuppDists version 1.1-9.9 Index]