ghyper.types {SuppDists}	R Documentation

Kemp and Kemp generalized hypergeometric types

Description

Generalized hypergeometric types as given by Kemp and Kemp

Two-way table

The basic representation is in terms of a two-way table:

x 	 k-x 	 k

a-x 	 b-k+x 	 N-k

a 	 b 	 N

and the associated hypergeometric probability P(x)=choose(a, x) choose(b, k-x) / choose(N, k).

The types are classified according to ranges of a, k, and N.

Kemp and Kemp types

Minor modifications in the definition of three of the types have been made to avoid numerical difficulties. Note, J denotes a nonnegative integer.

[Classic]

0<a, 0<N, 0<k

integers: a, N, k.

max(0,a+k-N) <= x <= min(a,k)

[IA(i)] (Real classic) 	 at least one noninteger parameter

0<a, 0<N, 0<k, k-1<a<N-(k-1)

integer: k

0 <= x <= k

[IA(ii)] (Real classic)	 at least one noninteger parameter

0<a, 0<N, 0<k, a-1<k<N-(a-1)

integer: a

0 <= x <= a

Interchanging a and k transforms this to type IA(i)

[IB]

0<a, 0<N, 0<k, a+k-1<N, J < (a,k) < J+1

integer: 0 <= J

non-integer: a, k

0 <= x …

NOTE: Kemp and Kemp specify -1<N.

No practical applications for this distribution.

[IIA] (negative hypergeometric)

a<0, N<a-1,0<k

integer: k

0 <= x <= k

NOTE: Kemp and Kemp specify N<a and N!=(a-1)

[IIB]

a<0, -1<N<k+a-1, 0<k, J < (k,k+a-1-N) < J+1

non-integer: k

integer: 0 <= J

0 <= x …

This is a very strange distribution. Special calculations were used.

Note: No practical applications.

[IIIA] (negative hypergeometric)

0<a,N<k-1,k<0

integer: a

0 <= x <= a

Interchanging a and k transforms this to type IIA

NOTE: Kemp and Kemp specify N<k and N != k-1

[IIIB]

0<a,-1<N<a+k-1,k<0, J<(a,a+k-1-N)<J+1

non integer: a

integer: 0 <= J

0 <= x …

Interchanging a and k transforms this to type IIB

Note: No practical applications

[IV] (Generalized Waring)

a<0,-1<N, k<0

0 <= x …

Author(s)

Bob Wheeler

References

Kemp, C.D., and Kemp, A.W. (1956). Generalized hypergeometric distributions. Jour. Roy. Statist. Soc. B. 18. 202-211. 39. 887-895.

[Package SuppDists version 1.1-9.9 Index]